pyproject.toml
[project]
name = "adknews"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "google-adk>=1.19.0",
    "litellm>=1.80.7",
    "opentelemetry-instrumentation-google-genai>=0.4b0",
]
